requirements.txt
# Core ML/AI Dependencies
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.30.0
datasets>=2.12.0
accelerate>=0.20.0
huggingface-hub>=0.15.0
einops>=0.6.0
bitsandbytes>=0.39.0

# Geometric Processing
open3d>=0.17.0
trimesh>=3.22.0
numpy>=1.24.0
scipy>=1.10.0
pyquaternion>=0.9.9

# CAD Processing
# Note: For non-pip installable packages like FreeCAD and OpenCascade,
# refer to README installation instructions
cadquery>=2.3.0
OCP>=7.7.0

# Computer Vision
opencv-python>=4.7.0
Pillow>=9.5.0
matplotlib>=3.7.0
imageio>=2.27.0

# Multi-modal Models
clip>=1.0
timm>=0.9.2

# Evaluation and Metrics
scikit-learn>=1.2.0
wandb>=0.15.0
tensorboard>=2.13.0
torchmetrics>=1.0.0
pynvml>=11.5.0
seaborn>=0.12.0

# Deployment
fastapi>=0.100.0
uvicorn>=0.22.0
pydantic>=2.0.0
python-multipart>=0.0.6
httpx>=0.24.0
gunicorn>=21.0.0

# Development Tools
pytest>=7.3.0
black>=23.3.0
flake8>=6.0.0
isort>=5.12.0
pre-commit>=3.3.0
jupyter>=1.0.0
ipywidgets>=8.0.0
mypy>=1.3.0
nbdime>=3.2.0

# CAD Rendering and Validation
meshio>=5.3.0
vtk>=9.2.0
pyrender>=0.1.45
pygltflib>=1.16.0

# Optimization and Training
optuna>=3.2.0
ray[tune]>=2.5.0
tqdm>=4.65.0
joblib>=1.2.0
psutil>=5.9.0

# Utilities
PyYAML>=6.0
requests>=2.30.0
click>=8.1.3
rich>=13.3.5
typer>=0.9.0
